Fixed folders names
The `data` property of FolderEntity is not Enumerable (anymore?) so while creating hashTable, the `data` property was not copied over. It's the issue because the script uses `folderEntity.data.name` later on to create directories named after folders. This allows to copy `data` property, so name is accessible later on. Now exporting on 0.8.9 version of FoundryVTT works.